+++ dev/children/vue2/src/pages/home.vue
@@ -23,12 +23,21 @@
       <button @click="renderNode">渲染一万个元素</button>
     </div>
     <router-link to="/page2">page2</router-link>
+
+    <div>{{drag?'拖拽中':'拖拽停止'}}</div>
+    <!--使用draggable组件-->
+    <draggable v-model="myArray"  chosenClass="chosen" forceFallback="true" group="people" animation="1000" @start="onStart" @end="onEnd">
+      <transition-group>
+      <div class="item" v-for="element in myArray" :key="element.id">{{element.name}}</div>
+      </transition-group>
+    </draggable>
   </div>
 </template>
 
 <script>
 import HelloWorld from '../components/HelloWorld.vue'
 import Vue from 'vue'
+import draggable from 'vuedraggable'
 
 export default {
   name: 'Page1',
@@ -38,6 +47,14 @@ export default {
       centerDialogVisible: false,
       microDataStr: '',
       liList: 0,
+      drag:false,
+      //定义要被拖拽对象的数组
+      myArray:[
+        {people:'cn',id:1,name:'www.itxst.com'},
+        {people:'cn',id:2,name:'www.baidu.com'},
+        {people:'cn',id:3,name:'www.taobao.com'},
+        {people:'us',id:4,name:'www.google.com'}
+      ]
     }
   },
   created () {
@@ -48,7 +65,8 @@ export default {
     window.microApp && window.microApp.removeDataListener(this.handleDataChange)
   },
   components: {
-    HelloWorld
+    HelloWorld,
+    draggable,
   },
   methods: {
     handleDataChange (data) {
@@ -72,7 +90,15 @@ export default {
     jumpWithLocation () {
       // 通过location跳转page2
       location.href = '/micro-app/vue2/#/page2'
-    }
+    },
+    //开始拖拽事件
+    onStart(){
+      this.drag=true;
+    },
+    //拖拽结束事件
+    onEnd() {
+      this.drag=false;
+    },
   }
 }
 </script>
@@ -98,5 +124,17 @@ export default {
   .test-safari-before::before {
     content: 'ddddd';
     color: red;
+  }
+  /*被拖拽对象的样式*/
+  .item {
+    padding: 6px;
+    background-color: #fdfdfd;
+    border: solid 1px #eee;
+    margin-bottom: 10px;
+    cursor: move;
+  }
+  /*选中样式*/
+  .chosen {
+    border: solid 2px #3089dc !important;
   }
 </style>