+++ dev/main-react16/src/pages/react16/react16.js
@@ -23,6 +23,8 @@ export default class App extends React.Component {
     showMicroApp: true,
     testNum: 0,
     showModal: false,
+    routerMode: 'custom',
+    baseroute: '/micro-app/demo/react16',
   }
 
   handleCreated = () => {
@@ -187,23 +189,34 @@ export default class App extends React.Component {
     })
   }
 
+  changeRouterMode = () => {
+    const newMode = {
+      custom: 'search',
+      search: 'history',
+      history: 'custom',
+    }
+
+    this.setState({
+      routerMode: newMode[this.state.routerMode],
+    })
+  }
+
   changeTestNum = () => {
     this.setState({
       testNum: this.state.testNum + 1,
     })
-    console.log(33333, this.props.history)
   }
 
   jumpToHome = () => {
-    microApp.router.push({name: this.state.name, path: '/micro-app/react16/'})
+    microApp.router.push({name: this.state.name, path: this.state.baseroute + '/'})
   }
 
   jumpToPage2 = () => {
-    microApp.router.push({name: this.state.name, path: '/micro-app/demo/react16/page2'})
+    microApp.router.push({name: this.state.name, path: this.state.baseroute + '/page2'})
   }
 
   jumpToInline = () => {
-    microApp.router.push({name: this.state.name, path: '/micro-app/react16/inline'})
+    microApp.router.push({name: this.state.name, path: this.state.baseroute + '/inline'})
   }
 
   useRouterGo = () => {
@@ -391,6 +404,7 @@ export default class App extends React.Component {
             {/* <Button type="primary" onClick={this.clearGlobalData}>清空全局数据</Button> */}
             <Button type="primary" onClick={this.changeNameUrl}>切换应用</Button>
             <Button type="primary" onClick={this.useUnmountApp}>主动卸载应用</Button>
+            <Button type="primary" onClick={this.changeRouterMode}>切换路由模式</Button>
             <Button type="primary" onClick={this.jumpToHome}>控制子应用跳转home</Button>
             <Button type="primary" onClick={this.jumpToPage2}>控制子应用跳转page2</Button>
             <Button type="primary" onClick={this.jumpToInline}>控制子应用跳转inline</Button>
@@ -422,7 +436,7 @@ export default class App extends React.Component {
                   onAftershow={this.handleAftershow}
                   onAfterhidden={this.handleAfterhidden}
                   onDataChange={this.handleDataChange}
-                  baseroute='/micro-app/demo/react16'
+                  baseroute={this.state.baseroute}
                   // keep-alive
                   // destroy
                   // inline