+++ src/sandbox/with/index.ts
@@ -11,7 +11,6 @@ import type {
 } from '@micro-app/types'
 import globalEnv from '../../libs/global_env'
 import microApp from '../../micro_app'
-import bindFunctionToRawTarget from '../bind_function'
 import {
   EventCenterForMicroApp,
   rebuildDataCenterSnapshot,
@@ -25,10 +24,8 @@ import {
   getEffectivePath,
   isArray,
   isPlainObject,
-  isString,
   isUndefined,
   removeDomScope,
-  unique,
   throttleDeferForSetAppName,
   rawDefineProperty,
   rawDefineProperties,
@@ -40,6 +37,7 @@ import {
   createProxyDocument,
 } from './document'
 import {
+  createProxyWindow,
   patchWindowEffect,
 } from './window'
 import {
@@ -95,18 +93,18 @@ export default class WithSandBox implements WithSandBoxInterface {
   private windowEffect: CommonEffectHook
   private documentEffect: CommonEffectHook
   private removeHistoryListener!: CallableFunction
-  private adapter: SandBoxAdapter
+  public adapter: SandBoxAdapter
   /**
    * Scoped global Properties(Properties that can only get and set in microAppWindow, will not escape to rawWindow)
    * Fix https://github.com/micro-zoe/micro-app/issues/234
    */
-  private scopeProperties: PropertyKey[] = []
+  public scopeProperties: PropertyKey[] = []
   // Properties that can be escape to rawWindow
-  private escapeProperties: PropertyKey[] = []
+  public escapeProperties: PropertyKey[] = []
   // Properties escape to rawWindow, cleared when unmount
-  private escapeKeys = new Set<PropertyKey>()
+  public escapeKeys = new Set<PropertyKey>()
   // Properties newly added to microAppWindow
-  private injectedKeys = new Set<PropertyKey>()
+  public injectedKeys = new Set<PropertyKey>()
   public proxyWindow!: proxyWindow // Proxy
   public microAppWindow = new EventTarget() as MicroAppWindowType // Proxy target
 
@@ -139,6 +137,7 @@ export default class WithSandBox implements WithSandBoxInterface {
     this.active = true
 
     /* --- memory router part --- start */
+    // create location, history for child app
     if (isUndefined(this.microAppWindow.location)) {
       this.setMicroAppRouter(
         this.microAppWindow.__MICRO_APP_NAME__,
@@ -147,6 +146,7 @@ export default class WithSandBox implements WithSandBoxInterface {
       )
     }
 
+    // update microLocation, attach route info to browser url
     this.initRouteState(defaultPage)
 
     // unique listener of popstate event for sub app
@@ -385,111 +385,6 @@ export default class WithSandBox implements WithSandBoxInterface {
     }
   }
 
-  // create proxyWindow with Proxy(microAppWindow)
-  private createProxyWindow (appName: string, microAppWindow: microAppWindowType) {
-    const rawWindow = globalEnv.rawWindow
-    const descriptorTargetMap = new Map<PropertyKey, 'target' | 'rawWindow'>()
-    return new Proxy(microAppWindow, {
-      get: (target: microAppWindowType, key: PropertyKey): unknown => {
-        throttleDeferForSetAppName(appName)
-        if (
-          Reflect.has(target, key) ||
-          (isString(key) && /^__MICRO_APP_/.test(key)) ||
-          this.scopeProperties.includes(key)
-        ) return Reflect.get(target, key)
-
-        return bindFunctionToRawTarget(Reflect.get(rawWindow, key), rawWindow)
-      },
-      set: (target: microAppWindowType, key: PropertyKey, value: unknown): boolean => {
-        /**
-         * TODO:
-         * 1、location域名相同，子应用内部跳转时的处理
-         */
-        if (this.adapter.escapeSetterKeyList.includes(key)) {
-          Reflect.set(rawWindow, key, value)
-        } else if (
-          // target.hasOwnProperty has been rewritten
-          !rawHasOwnProperty.call(target, key) &&
-          rawHasOwnProperty.call(rawWindow, key) &&
-          !this.scopeProperties.includes(key)
-        ) {
-          const descriptor = Object.getOwnPropertyDescriptor(rawWindow, key)
-          const { configurable, enumerable, writable, set } = descriptor!
-          // set value because it can be set
-          rawDefineProperty(target, key, {
-            value,
-            configurable,
-            enumerable,
-            writable: writable ?? !!set,
-          })
-
-          this.injectedKeys.add(key)
-        } else {
-          !Reflect.has(target, key) && this.injectedKeys.add(key)
-          Reflect.set(target, key, value)
-        }
-
-        if (
-          (
-            this.escapeProperties.includes(key) ||
-            (
-              this.adapter.staticEscapeProperties.includes(key) &&
-              !Reflect.has(rawWindow, key)
-            )
-          ) &&
-          !this.scopeProperties.includes(key)
-        ) {
-          !Reflect.has(rawWindow, key) && this.escapeKeys.add(key)
-          Reflect.set(rawWindow, key, value)
-        }
-
-        return true
-      },
-      has: (target: microAppWindowType, key: PropertyKey): boolean => {
-        if (this.scopeProperties.includes(key)) return key in target
-        return key in target || key in rawWindow
-      },
-      // Object.getOwnPropertyDescriptor(window, key)
-      getOwnPropertyDescriptor: (target: microAppWindowType, key: PropertyKey): PropertyDescriptor|undefined => {
-        if (rawHasOwnProperty.call(target, key)) {
-          descriptorTargetMap.set(key, 'target')
-          return Object.getOwnPropertyDescriptor(target, key)
-        }
-
-        if (rawHasOwnProperty.call(rawWindow, key)) {
-          descriptorTargetMap.set(key, 'rawWindow')
-          const descriptor = Object.getOwnPropertyDescriptor(rawWindow, key)
-          if (descriptor && !descriptor.configurable) {
-            descriptor.configurable = true
-          }
-          return descriptor
-        }
-
-        return undefined
-      },
-      // Object.defineProperty(window, key, Descriptor)
-      defineProperty: (target: microAppWindowType, key: PropertyKey, value: PropertyDescriptor): boolean => {
-        const from = descriptorTargetMap.get(key)
-        if (from === 'rawWindow') {
-          return Reflect.defineProperty(rawWindow, key, value)
-        }
-        return Reflect.defineProperty(target, key, value)
-      },
-      // Object.getOwnPropertyNames(window)
-      ownKeys: (target: microAppWindowType): Array<string | symbol> => {
-        return unique(Reflect.ownKeys(rawWindow).concat(Reflect.ownKeys(target)))
-      },
-      deleteProperty: (target: microAppWindowType, key: PropertyKey): boolean => {
-        if (rawHasOwnProperty.call(target, key)) {
-          this.injectedKeys.has(key) && this.injectedKeys.delete(key)
-          this.escapeKeys.has(key) && Reflect.deleteProperty(rawWindow, key)
-          return Reflect.deleteProperty(target, key)
-        }
-        return true
-      },
-    })
-  }
-
   /**
    * create proxyWindow, rewrite window event & timer of child app
    * @param appName app name
@@ -497,7 +392,7 @@ export default class WithSandBox implements WithSandBoxInterface {
    */
   private patchWithWindow (appName: string, microAppWindow: microAppWindowType): CommonEffectHook {
     // create proxyWindow with Proxy(microAppWindow)
-    this.proxyWindow = this.createProxyWindow(appName, microAppWindow)
+    this.proxyWindow = createProxyWindow(appName, microAppWindow, this)
     // rewrite global event & timeout of window
     return patchWindowEffect(appName, microAppWindow)
   }